Move AuthorityClient to a separate file
